Remove deadcode in context_groups view
This code was being investigated as part of the 5VPAT work.
Based off of the investigation, it seems that this code is
unreachable and no longer used. If no tests fail, then it
may be safe to assume this is deadcode and we can remove it.

closes EGG-409
flag=deprecate_context_groups_old_view

Test Plan:
- Jenkins build passes
- With the FF enabled, verify that groups page renders
  successfully as a teacher and a student

[skip-crystalball]

Change-Id: Id10eb22a9ba9180096194336e5aa4640b794337c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/368997
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jason Gillett <jason.gillett@instructure.com>
QA-Review: Jason Gillett <jason.gillett@instructure.com>
Product-Review: Kyle Rosenbaum <krosenbaum@instructure.com>