Support upload zip files and name conflicts
closes RCX-2819
closes RCX-2820

flag=files_a11y_rewrite

test plan ZIP:
- Navigate to a course/user files page.
- Click Upload button and select a zip file.
> Verify that a modal should show  the
buttons Expand it and Upload it.
> If you click Expand button, verify that the
upload is performed and new folder with
the zip content is created.
> If you click Upload button, verify that the
upload is performed and the file appears
in the current folder.

test plan RENAME:
- Navigate to a course/user files page.
- Click Upload button and select a regular file.
- Try to upload another file with the same name.
> Verify that a modal should show the
buttons Skip, Change name and Replace.
> If you click Skip button, verify that the
the upload is not performed.
> If you click Replace button, verify that the
the upload is performed and the file was
replaced in the current folder.
> If you click Change name button, verify that
the modal body changes and an text input appears.
> If you click Back, verify the modal body is
restored with the initial content.
> If there is still a name conflict and click Change,
verify that it should return to the original
modal.
> If the new name also has conflicts and click Change,
nothing happens and the modal flow  is repeated
but with the new name until there are no name
conflicts.

NOTE: The two modals could appear on the same
flow, for example after uploading a zip with the
name conflicts.

Change-Id: I373c82ce36c821e9c935bb8a4ed4b8eff1d86150
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/369423
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jacob DeWar <jacob.dewar@instructure.com>
QA-Review: Jacob DeWar <jacob.dewar@instructure.com>
Product-Review: Juan Chavez <juan.chavez@instructure.com>