Upgrade Canvas to InstUI version 10
Refs INSTUI-4364
Closes FOO-5166
flag=none
[skip-crystalball]
[fsc-max-nodes=25]
[fsc-timeout=60]

This upgrades Canvas-LMS to InstUI version 10. Major changes
to theming and component appearance were made, so this will
require a thorough manual once-over to verify that everything
still looks acceptable product-wide. With a fully passing
test suite, functionality is not likely to be impacted. Most
of the changes in this commit were to fix tests and to update
theme overrides to the new color naming scheme in InstUI 10.

Please do not merge until we determine a proper testing
schedule for this.

Test plan:
* This was a major UI change, so all Canvas teams should
  carefully and fully inspect all the features they are
  responsible for to verify that they still look and
  function as desired

Change-Id: I4ddc8686c1b6fcf29717669b31dbef31ff114b53
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/369859
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Maya Tyner <maya.tyner@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>