Fully Typescriptify ui/features/release_notes_edit
Refs FOO-4955
flag=none

Working on the CreateEditModal, enough work was done on
typing sanity that it made some sense to just keep pushing
through a full conversion of the entire directory in order
to replace all JavaScript to TypeScript. I'm pretty happy
with the result, it caught a couple of actual errors in
the code.

This also arranges to render the release notes editor
via React Router instead of a js_bundle.

This also spiffs up the interfaces to doFetchApi and useFetchApi
to make explicit that the Error objects they return have been
augmented with the Response object from fetch(). It's cleaner
and also makes TypeScript a little happier.

Test plan:
* Automated tests pass
* Release note editing feature works as before

Change-Id: I93be639f488014a5b9197e4d84b85f34ae4d20b8
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/370328
Reviewed-by: Adrián Grúber <adrian.gruber@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>