allow ordering LTI ToolFinder by context chain
why
* some callers requesting "all tools available in this context"
are wanting those tools nearest in the context chain to be
returned first
* centralize this ordering in the ToolFinder
* the ContextToolFinder now takes keyword args instead of
an options hash

closes INTEROP-9214
flag=none

test plan:
* specs pass everywhere
* install an LTI tool in multiple places - in the root account,
maybe in a subaccount, and in a course. it doesn't matter whether
this is a 1.1 or 1.3 tool. this can also be different LTI tools, as long
as there are tools at each context level
* in a Rails console, run
```
Lti::ContextToolFinder.ordered_by_context_for(course)
```
* unlike `.all_tools_for`, the tools installed in the course should be
first in this list, and the tools installed at the root account should
be last

Change-Id: I87a1f22fb7a4c5fc1ba4f4841f9e41803c78c0a6
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/372713
Reviewed-by: Tucker Mcknight <tmcknight@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Xander Moffatt <xmoffatt@instructure.com>