Remove aria-hidden from wiki_page_body textarea
Chrome and Firefox are handling this situation differently and Chrome
explains why. In Chrome, when you switch to the raw HTML editor it puts
this error in the console: "Blocked aria-hidden on an element because its
descendant retained focus. The focus must not be hidden from assistive
technology users. Avoid using aria-hidden on a focused element or its
ancestor." This means Chrome is ignoring aria-hidden and telling us about
it. Firefox is not ignoring it and is respecting the hidden. This is why
screenreaders on Chrome are working and Firefox are not.

The fix here is to remove that aria-hidden attribute to make all browsers
happy with the behavior we want.

fixes RCX-2491
flag=none

test plan:
- Open the RCE in Firefox with a screenreader on
- Open the raw HTML editor
- Verify the screenreader reads content inside the raw HTML editor
- Verify the same happens in Chrome/Safari/etc.

Change-Id: I88cf263de4cd6899d94da5d27dfdca5dd7a75231
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/372917
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Sarah Gerard <sarah.gerard@instructure.com>
QA-Review: Sarah Gerard <sarah.gerard@instructure.com>
Product-Review: Kata Virag Keri <kata.keri@instructure.com>