Fix courses index page sort icons contrast
fixes LMA-36

flag=none

test plan:
- As a user with at least one enrollment, navigate
to all courses.
> Verify that the icons constrasts with the
background.
NOTE: I noticed that after darken the original color,
the user couldn't notice the difference when the sort
is applied or not. So based on what we have in INST-UI,
I'm using a contrasted gray in case it's not applied
and a constrasted blue when the sorting is applied.

Change-Id: Id834c75b5161264b509acc4ae11d8e08a3b361ff
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/375187
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Luis Hilario <luis.hilario@instructure.com>
QA-Review: Luis Hilario <luis.hilario@instructure.com>
Product-Review: Peter Budavari <peter.budavari@instructure.com>