remove remaining authlogic basic auth support code
we removed / disabled basic auth support back in 2015, but it looks
like we still had some code in the initializers that was still
supporting the use of it.  This commit removes the last of that code.

closes FOO-
flag = none

test plan:
 - our usage of authlogic sessions for auth should not be affected

Change-Id: I753113bf9f82dbb12dee9b20bac0128120df6f4a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/375204
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>