Fix bad behavior in Course Settings default due time
Refs FOO-823
Fixes FOO-5341
flag=none

The CourseDefaultDueTime component was written to replace
the older jQuery picker, but it had a lot of issues. It would
show the current time instead of the account default (or
whatever previous value was set), and would often not save
a value when changed.

This commit more or less rewrote that component so that it
initializes and then maintains the proper state, uses the
InstUI <TimeSelect> component in a controlled way, and does
a better job of validating input. A few more tests were added
as well, to check this functionality.

This also improves how the default due time is created on
the back end to more consistently reflect the notion of
"fancy midnight" i.e. 11:59:59pm.

All of this is still behind a feature flag and not in Production
and actually may never be, but this improves code quality
in any event. 🤷

Test plan:
* Have a course in an account and visit its setting page
* The "Default due time" should match the acccount default
  (probably 11:59pm or 23:59 depending on your locale) and
  should respond to changes either by picking off the popup
  menu of half-hour choices or manually typing in a value
* If you type something illegal (i.e. that doesn't parse to
  a time value) and then blur out of the field, it should
  reset to whatever last legal value was there
* After making a change, submitting the form via "Update
  Course Details" should persist the change

Change-Id: I9fc52119e1eff158fcc63c61468788f27b71bcf1
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/376133
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>