don't "temporarily" change files' content_type
this hack was implemented in 2017 to prevent a file
from being linked to an existing root attachment
when we're trying to make it independent

but there's a terrible bug in the logic: it does not
anticipate md5 being nil. also, it does not actually
verify that other files matching the same md5 have
the same content_type, and just overwrites them all.

but also, it's a silly hack to begin with. if we want to
prevent find_existing_attachment_for_md5 from finding
other files, we can just _ask it not to_.

test plan: with local storage,
 - upload the same file twice
 - verify the second file points at the first via
   root_attachment_id
 - call `make_rootless` on the second file
 - verify its root_attachment_id is nil
   and its filename points at an independent copy
   in the file store

flag=none
closes FOO-5348

Change-Id: I7a6518a01cbad63d8364dd1437fd1fd5592b7d0b
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/375542
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Mysti Lilla <mysti@instructure.com>
QA-Review: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Jeremy Stanley <jeremy@instructure.com>