memoize docviewer feature flag hook
Looking up if docviewer can use iwork or not is expensive. If things are
repeatedly checking this flag in a loop it can be a performance
bottleneck. Further, this data does not change often, so we can afford
to let an app server hang on to it for a bit.  This memoizes the result
of the check per context

flag=none

test plan:
 - specs pass

Change-Id: I858312f307339cba10d099e36132c358350cf84a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/375877
Reviewed-by: Spencer Olson <solson@instructure.com>
Reviewed-by: Kasra Rismanchi <kasra.rismanchi@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Keith Garner <kgarner@instructure.com>
Product-Review: Keith Garner <kgarner@instructure.com>