Handle file collisions
due to pagination in the new files UI, if an uploaded
file conflicts with any files that are not on the current
page, then we are unable to catch the conflict. This changes
the workflow so we attempt to upload the conflicting file
and then catch the error from the api, then allow the user
to try and resolve the conflict.
If there is a conflict on the current page, then we still try
to catch and resolve that conflict before we send off the request.

closes RCX-3256
flag=files_a11y_rewrite

test plan:
- have enough files in a folder to have pagination (25+)
- from the first page, attempt to upload a file which has a
  conflicting name with a file on the second page
- expect the file upload ui to be triggered, then the rename
  modal to appear.
- expect all three of the modal options to be functional, as well
  as the cancel button
- attempt with multiple uploads at once/multiple conflicts
- attempt to change the name of a file to another conflicting file,
  expect the modal to resolve the conflict to appear again.
- If you upload a file which conflicts with a file on the current
  page, expect the conflict to be caught and resolve before the
  upload UI is triggered.

Change-Id: I600dc035200daa429fe6b7e1e15ad6bcd5486b00
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/376101
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Eric Saupe <eric.saupe@instructure.com>
QA-Review: Eric Saupe <eric.saupe@instructure.com>
Product-Review: Sarah Gerard <sarah.gerard@instructure.com>